refactor address

+++ smart-rent/src/main/java/com/smartrent/dto/response/ListingResponseWithAdmin.java
@@ -68,7 +68,7 @@ public class ListingResponseWithAdmin {
 
     String priceUnit;
 
-    Long addressId;
+    Integer addressId;
 
     Float area;
